+++ components/SignPages/SignHeader/styles.module.css
@@ -1,0 +1,18 @@
+.header {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 16px;
+}
+
+.header__text {
+  font-size: 1.6rem;
+  line-height: 24px;
+}
+
+.header__sign-link {
+  font-weight: 600;
+  line-height: normal;
+  color: var(--Linkbrary-primary-color);
+  text-decoration: underline;
+}
