--- styles/signup.module.css
@@ -1,36 +1,0 @@
-.background-container {
-  background-color: var(--Linkbrary-bg);
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-}
-
-.sign-section {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 400px;
-  padding: 238px 0 100%;
-  row-gap: 30px;
-}
-
-.header {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 16px;
-}
-
-.header__text {
-  font-size: 1.6rem;
-  line-height: 24px;
-}
-
-.header__signin-link {
-  font-weight: 600;
-  line-height: normal;
-  color: var(--Linkbrary-primary-color);
-  text-decoration: underline;
-}
